requirements.txt
Flask
Flask-SQLAlchemy
gunicorn
openpyxl
python-dotenv
psycopg2-binary